+++ packages/data/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./opportunity-index";
 export * from "./loaders";
+export * from "./validation";